# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API documentation for Laravel project",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Laravel API Server"
        }
    ],
    "paths": {
...
        "/api/account/cancel-deletion": {
            "post": {
                "tags": [
                    "Account Deletion"
                ],
                "summary": "Cancel pending deletion request",
                "description": "Cancels the 7-day scheduled deletion process.",
                "operationId": "f17d2cce08d2d8419e4a83fc37f65a00",
                "responses": {
                    "200": {
                        "description": "Successfully canceled account deletion"
                    },
                    "400": {
                        "description": "No pending deletion request found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Creates a new user account and returns authentication token with profile data",
                "operationId": "d6f5d00c40f9de2e9272088aebe648c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "password",
                                    "auth_type",
                                    "firebase_id"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "password123"
                                    },
                                    "auth_type": {
                                        "type": "string",
                                        "enum": [
                                            "gmail",
                                            "apple",
                                            "email_pass"
                                        ],
                                        "example": "email_pass"
                                    },
                                    "firebase_id": {
                                        "type": "string",
                                        "example": "abcd343kejfrn"
                                    },
                                    "phone_country_code": {
                                        "type": "string",
                                        "maxLength": 5,
                                        "example": "+1"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "1234567890"
                                    },
                                    "profile_img": {
                                        "description": "Base64 encoded image or URL",
                                        "type": "string",
                                        "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQ...",
                                        "nullable": true
                                    },
                                    "app_version": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "1.0.0"
                                    },
                                    "ip_address": {
                                        "type": "string",
                                        "format": "ipv4",
                                        "example": "192.168.1.1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Account created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abcdef1234567890"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "example": "60839236"
                                                },
                                                "profile": {
                                                    "properties": {
                                                        "user_id": {
                                                            "type": "string",
                                                            "example": "60839236"
                                                        },
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "john.doe@example.com"
                                                        },
                                                        "phone": {
                                                            "properties": {
                                                                "country_code": {
                                                                    "type": "string",
                                                                    "example": "+1"
                                                                },
                                                                "number": {
                                                                    "type": "string",
                                                                    "example": "1234567890"
                                                                }
                                                            },
                                                            "type": "object",
                                                            "nullable": true
                                                        },
                                                        "profile_img": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "auth_type": {
                                                            "type": "string",
                                                            "example": "email_pass"
                                                        },
                                                        "app_version": {
                                                            "type": "string",
                                                            "example": "1.0.0",
                                                            "nullable": true
                                                        },
                                                        "acc_status": {
                                                            "type": "string",
                                                            "example": "active"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Registration failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registration failed: Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "description": "Authenticate user with email and password, return access token",
                "operationId": "0f5784f47a3fcac35f86ec4e82d967d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sign in successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "2|ghijkl1234567890"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "example": "60839236"
                                                },
                                                "profile": {
                                                    "properties": {
                                                        "user_id": {
                                                            "type": "string",
                                                            "example": "60839236"
                                                        },
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "john.doe@example.com"
                                                        },
                                                        "phone": {
                                                            "type": "object",
                                                            "nullable": true
                                                        },
                                                        "profile_img": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "auth_type": {
                                                            "type": "string",
                                                            "example": "email_pass"
                                                        },
                                                        "app_version": {
                                                            "type": "string",
                                                            "nullable": true
                                                        },
                                                        "acc_status": {
                                                            "type": "string",
                                                            "example": "active"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Login failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login failed: Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "description": "Revoke current access token and logout user",
                "operationId": "d07352aa1a19c3f27d2df5ced5f833cf",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Logout failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to logout: Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout-all": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from all devices",
                "description": "Revoke all access tokens for the user from all devices",
                "operationId": "9d84edd110d07626134670dc23f75075",
                "responses": {
                    "200": {
                        "description": "Logout from all devices successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out from all devices."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Logout failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to logout from all devices: Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}